Refactor books options
