+++ app.rb
@@ -1,11 +1,11 @@
-require_relative './people_options.rb'
-require_relative './book'
+require_relative './people_options'
+require_relative './books_options'
 require_relative './rental'
 
 class App
   def initialize(menu)
     @menu = menu
-    @books = []
+    @books_options = BooksOptions.new
     @people_options = PeopleOptions.new
     @rentals = []
   end
@@ -13,41 +13,22 @@ class App
   def select_option(user_input)
     case user_input
     when 1
-      list_all_books
+      @books_options.list_all_books
     when 2
       @people_options.list_all_people
     when 3
       @people_options.create_person
     when 4
-      create_book
+      @books_options.create_book
     when 5
       create_rental
     when 6
       list_all_rentals
     when 7
-      exit
+      exit_app
     else
       puts 'Please select the number from the given list!'
     end
-    @menu.show_menu
-  end
-
-  def list_all_books
-    if @books.empty?
-      puts 'No added book!'
-    else
-      @books.each { |book| puts "Title: #{book.title}, Author: #{book.author}" }
-    end
-    @menu.show_menu
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    @books.push(Book.new(title, author))
-    puts 'Book created successfully'
     @menu.show_menu
   end
 
@@ -81,7 +62,8 @@ class App
     @menu.show_menu
   end
 
-  def exit
+  def exit_app
     puts 'Thank you for using the app!'
+    exit(true)
   end
 end
